backend/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "Backend for the Phase-2 Web-Based Todo Application"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy>=2.0.35",
    "sqlalchemy-utils>=0.41.1",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "alembic>=1.13.3",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.8.0",
    "flake8>=7.1.0",
]